convert getErrorAction to notificationError middleware
